+++ sql/dijkstra/_dijkstraVia.sql
@@ -43,7 +43,7 @@ CREATE FUNCTION _pgr_dijkstraVia(
     OUT path_seq INTEGER,
     OUT start_vid BIGINT,
     OUT end_vid BIGINT,
-    OUT prev BIGINT,
+    OUT pred BIGINT,
     OUT node BIGINT,
     OUT edge BIGINT,
     OUT cost FLOAT,