+++ sql/trsp/_trspVia.sql
@@ -40,7 +40,7 @@ CREATE FUNCTION _pgr_trspVia(
   OUT path_seq INTEGER,
   OUT start_vid BIGINT,
   OUT end_vid BIGINT,
-  OUT prev BIGINT,
+  OUT pred BIGINT,
   OUT node BIGINT,
   OUT edge BIGINT,
   OUT cost FLOAT,