+++ sql/trsp/trspVia.sql
@@ -37,7 +37,7 @@ CREATE FUNCTION pgr_trspVia(
   OUT path_seq INTEGER,
   OUT start_vid BIGINT,
   OUT end_vid BIGINT,
-  OUT prev BIGINT,
+  OUT pred BIGINT,
   OUT node BIGINT,
   OUT edge BIGINT,
   OUT cost FLOAT,
@@ -45,7 +45,7 @@ CREATE FUNCTION pgr_trspVia(
   OUT route_agg_cost FLOAT)
 RETURNS SETOF RECORD AS
 $BODY$
-SELECT seq, path_id, path_seq, start_vid, end_vid, prev, node, edge, cost, agg_cost, route_agg_cost
+SELECT seq, path_id, path_seq, start_vid, end_vid, pred, node, edge, cost, agg_cost, route_agg_cost
 FROM _pgr_trspVia(
   _pgr_get_statement($1),
   _pgr_get_statement($2),