+++ src/cpp_common/to_postgres.cpp
@@ -58,7 +58,7 @@ get_path(
         double &route_agg_cost,
         size_t &sequence) {
     int path_seq = 0;
-    int64_t prev_node = path.start_id();
+    int64_t pred = path.start_id();
     for (const auto e : path) {
         (*tuples)[sequence] = {
             route_id,
@@ -66,14 +66,14 @@ get_path(
             path_seq,
             path.start_id(),
             path.end_id(),
-            prev_node,
+            pred,
             e.node,
             e.edge,
             e.cost,
             e.agg_cost,
             route_agg_cost};
         route_agg_cost += path[static_cast<uint32_t>(path_seq)].cost;
-        prev_node = e.node;
+        pred = e.node;
         path_seq++;
         ++sequence;
     }