+++ src/coloring/coloring_driver.cpp
@@ -35,6 +35,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <string>
 #include <utility>
 #include <cstdint>
+#include <algorithm>
 
 #include "c_types/ii_t_rt.h"
 #include "cpp_common/base_graph.hpp"
@@ -44,6 +45,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "cpp_common/alloc.hpp"
 #include "cpp_common/assert.hpp"
 
+#include "planar/makeBiconnectedPlanar.hpp"
 #include "coloring/bipartite.hpp"
 #include "coloring/edgeColoring.hpp"
 #include "coloring/sequentialVertexColoring.hpp"
@@ -121,11 +123,20 @@ void do_coloring(
         } else {
             if (which == CORENUMBERS) {
                 undigraph.insert_cost1_edge_no_parallel_no_loop(edges);
+            } else if (which == BICONNECTEDPLANAR) {
+                undigraph.insert_cost1_edge_no_parallel_no_loop(edges);
             } else {
                 undigraph.insert_edges(edges);
             }
 
             switch (which) {
+               case BICONNECTEDPLANAR:
+                    {
+                        pgrouting::functions::Pgr_makeBiconnectedPlanar<UndirectedGraph> fn;
+                        results = fn.makeBiconnectedPlanar(undigraph);
+                        log << fn.get_log();
+                    }
+                    break;
                 case EDGECOLORING:
                     results = edgeColoring(undigraph);
                     break;