+++ doc/src/release_notes.rst
@@ -50,7 +50,7 @@ To see all issues & pull requests closed by this release see the
 * :issue:`3089`: edgeDisjoint and bellmanFord use shortestPath driver and
   process
 * :issue:`3100`: Coloring: create and use a process & driver
-* :issue:`3111`: Via functions: add prev result column
+* :issue:`3111`: Via functions: add pred result column
 * :issue:`3113`: Components: Integrate into existing process/driver pair
 * :issue:`3129`: breadthFirstSearch: Reorganize into traversal
 * :issue:`3131`: binaryBreadthFirstSearch: Integrate into existing